Single file fuzzoutstation.cpp.
